requirements.txt
Django==1.8.19
Pillow==5.2.0
